refactor: conversion into ts index root-store
