+++ packages/cfd/src/Stores/Modules/index.ts
@@ -1,7 +1,10 @@
 import CFDStore from './CFD/cfd-store';
+import { TRootStore } from '../../types';
 
 export default class ModulesStore {
-    constructor(root_store) {
+    cfd;
+
+    constructor(root_store: TRootStore) {
         this.cfd = new CFDStore({ root_store });
     }
 }
