+++ packages/cfd/src/types/root-store.types.ts
@@ -1,0 +1,11 @@
+import type { TStores } from '@deriv/stores/types';
+import type CFDStore from '../Stores/Modules/CFD/cfd-store';
+
+/**
+ * @deprecated - Use `TStores` from `@deriv/stores` instead of this type.
+ */
+export type TRootStore = TStores & {
+    modules: {
+        cfd: CFDStore;
+    };
+};
